Reviewed-on: https://git.zach-dev.cc/zach/z-bar-qt/pulls/132

+++ Plugins/ZShell/Internal/stroke.hpp
@@ -11,9 +11,9 @@ struct Stroke {
 	QVector<QPointF> points;
 	QCanvasPath path;
 	QColor color;
-	float width;
+	qreal width;
 	int groupId = -1;
 	bool isSinglePoint = false;
 };
 
-};
+}; // namespace ZShell::internal
